// MumbaiHacks-server/src/Interfaces/income.interface.ts
import { ObjectId } from "mongoose";
import { Document } from "mongoose";

export interface IIncome extends Document {
    organizationId: ObjectId;
    amount: number;
    description?: string;
    date: Date;
    paymentMethod: 'Cash' | 'Credit Card' | 'Bank Transfer';
    tags?: string[];
    createdBy: ObjectId;
    payeeType: 'User' | 'Organization';
    payeeId: ObjectId;
}